タイマー作成と編集画面の共通メソッドをプロトコルエクステンションに追加
